Computer World - Small tweaks
Split up code into separate files.
Renamed database folder to db

Need to implement the biggest piece of feedback: Put validation first and then do the business code. Will hopefully finish this by tomorrow evening.
