+++ db/model/manufacturer.go
@@ -1,0 +1,6 @@
+package model
+
+type Manufacturer struct {
+	ManufacturerID   int    `gorm:"autoIncrement;primaryKey" json:"manufacturer_id"`
+	ManufacturerName string `gorm:"unique" json:"manufacturer_name"`
+}
