+++ db/model/product.go
@@ -1,0 +1,11 @@
+package model
+
+type Product struct {
+	ProductID      int    `gorm:"primaryKey;autoIncrement" json:"product_id"`
+	ProductCode    string `gorm:"unique" json:"product_code"`
+	ProductName    string `gorm:"unique" json:"product_name"`
+	ManufacturerID int    `json:"manufacturer_id"`
+	Manufacturer   Manufacturer
+	Stock          int     `json:"stock"`
+	Price          float64 `gorm:"not null" json:"price"`
+}
