+++ db/seed/seeder.go
@@ -1,0 +1,11 @@
+package seed
+
+import "gorm.io/gorm"
+
+type Seeding struct {
+	database *gorm.DB
+}
+
+func NewSeed(db *gorm.DB) *Seeding {
+	return &Seeding{database: db}
+}
