--- server/reponses/responses.go
@@ -1,8 +1,0 @@
-package reponses
-
-import "github.com/labstack/echo/v4"
-
-func ErrorResponse(context echo.Context, code int, err error) error {
-	context.Response().Status = code
-	return echo.NewHTTPError(code, err.Error())
-}
