+++ server/requests/manufacturer_requests.go
@@ -1,0 +1,5 @@
+package requests
+
+type ManufacturerRequest struct {
+	ManufacturerName string `json:"manufacturer_name"`
+}
