+++ server/requests/order_requests.go
@@ -1,0 +1,8 @@
+package requests
+
+type OrderRequest struct {
+	OrderReference string  `json:"order_reference"`
+	OrderAmount    int     `json:"order_amount"`
+	ProductID      int     `json:"product_id"`
+	ProductPrice   float64 `json:"product_price"`
+}
